requirements.txt
requests==2.31.0
sendgrid==6.11.0
django-sendgrid-v5==1.2.3
 